% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcensReg.R
\name{cens_trunc_param_check}
\alias{cens_trunc_param_check}
\title{Checking the censored and truncation values}
\usage{
cens_trunc_param_check(y, a, v, xi, b)
}
\arguments{
\item{y}{Observed outcome}

\item{a}{Left-truncation value}

\item{v}{Left-censoring value}

\item{xi}{Right-censoring value}

\item{b}{Right-truncation value}
}
\description{
Internal helper function to perform basic parameter checks
}
\keyword{internal}
