% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.tcensReg.R
\name{coef.tcensReg}
\alias{coef.tcensReg}
\title{Method for Extracting Coefficient Estimates for tcensReg S3 Objects}
\usage{
\method{coef}{tcensReg}(object, logsigma = FALSE, digits = 4, ...)
}
\arguments{
\item{object}{Object of class \code{tcensReg}}

\item{logsigma}{Logical indicator of whether to return standard deviation as logsigma or sigma. Default is FALSE returning sigma}

\item{digits}{Numeric scalar indicating the number of digits to return}

\item{...}{Extra arguments}
}
\value{
numeric named vector of coefficient estimates
}
\description{
Method for Extracting Coefficient Estimates for tcensReg S3 Objects
}
\keyword{internal}
