% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.tcensReg.R
\name{predict.tcensReg}
\alias{predict.tcensReg}
\title{Method for Obtaining Fitted Predictions for tcensReg S3 Objects}
\usage{
\method{predict}{tcensReg}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{Object of class \code{tcensReg}}

\item{newdata}{New values to obtain predictions}
}
\value{
numeric vector of fitted values, i.e., yhat
}
\description{
Method for Obtaining Fitted Predictions for tcensReg S3 Objects
}
\keyword{internal}
