% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tcensReg.R
\name{print.tcensReg}
\alias{print.tcensReg}
\title{Method for Printing tcensReg S3 Objects}
\usage{
\method{print}{tcensReg}(x, logsigma = FALSE, digits = 4, ...)
}
\arguments{
\item{x}{Object of class \code{tcensReg}}

\item{logsigma}{Logical indicator of whether to return standard deviation as logsigma or sigma. Default is FALSE returning sigma}

\item{digits}{Numeric scalar indicating the number of digits to return}

\item{...}{Additional optional arguments to pass into the printing process}
}
\value{
default method for printing an S3 object of class tcensReg
}
\description{
Method for Printing tcensReg S3 Objects
}
\keyword{internal}
