% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psuedo_r2.R
\name{psuedo_r2}
\alias{psuedo_r2}
\title{Regression Method for Truncated Normal Distribution with Censored Data
#'}
\usage{
psuedo_r2(obj, type = c("mckelvey_zavoina"))
}
\arguments{
\item{obj}{Object of class \code{formula} which symbolically describes the model to be fit}

\item{type}{Character value indicating the type of psuedo R^2 to calculate. Currently only mckelvey_zavoina is available}
}
\value{
List with numeric value representing the psuedo R^2 and type of psuedo R^2 calculcated
}
\description{
Regression Method for Truncated Normal Distribution with Censored Data
#'
}
\examples{
#truncated normal underlying data
y_star <- rtnorm(n = 1000, mu = 0.5, sd = 1, a = 0)

#apply censoring
y <- ifelse(y_star <= 0.25, 0.25, y_star)

#find MLE estimates
mod_result <- tcensReg(y ~ 1, v = 0.25, a = 0)

psuedo_r2(mod_result, type="m")

}
