% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tcensReg.R
\name{summary.tcensReg}
\alias{summary.tcensReg}
\title{Method for Returning Summary Statistics for tcensReg S3 Objects}
\usage{
\method{summary}{tcensReg}(object, logsigma = FALSE, digits = 4, ...)
}
\arguments{
\item{object}{Object of class \code{tcensReg}}

\item{logsigma}{Logical indicator of whether to return standard deviation as logsigma or sigma. Default is FALSE returning sigma}

\item{digits}{Numeric value for number of digits to display for all numeric results}

\item{...}{Additional optional parameters}
}
\value{
Relevant results of tcensReg fitted object
}
\description{
Method for Returning Summary Statistics for tcensReg S3 Objects
}
\keyword{internal}
