% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcensReg_gradient.R
\name{tcensReg_gradient}
\alias{tcensReg_gradient}
\title{Gradient Vector for Truncated Normal Distribution with Censoring with Linear Equation Mean}
\usage{
tcensReg_gradient(theta, y, X, a = -Inf, v = NULL)
}
\arguments{
\item{theta}{Numeric vector containing estimates of \eqn{\beta} and \eqn{\log(\sigma)}}

\item{y}{Numeric vector with the observed truncated and censored outcomes}

\item{X}{Numeric design matrix}

\item{a}{Numeric scalar indicating the truncation value. Initial value is -Inf indicating no truncation}

\item{v}{Numeric scalar indicating the censoring value. Initially set to NULL indicating no censoring}
}
\value{
Vector of gradient values with p-1 beta parameters and log sigma for the nth iterate
}
\description{
Gradient Vector for Truncated Normal Distribution with Censoring with Linear Equation Mean
}
\keyword{internal}
