% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcensReg_gradient_sepvar_maxLik.R
\name{tcensReg_gradient_sepvar_maxLik}
\alias{tcensReg_gradient_sepvar_maxLik}
\title{Gradient Vector for Truncated Normal Distribution with Censoring with Linear Equation Mean for J Independent Truncated Normals with Seprate Variance}
\usage{
tcensReg_gradient_sepvar_maxLik(theta, y, X, group, left_trunc = -Inf,
  v = NULL)
}
\arguments{
\item{theta}{Numeric vector numeric vector containing estimates of beta and log sigma}

\item{y}{Numeric vector with the observed truncated and censored outcomes}

\item{X}{Numeric design matrix}

\item{group}{Character vector identifying the group membership for the independent truncated normal variables. This defines the \code{J} groups.}

\item{left_trunc}{Numeric scalar indicating the truncation value}

\item{v}{Numeric scalar indicating the censoring value}
}
\value{
Vector of gradient values with p-1 beta parameters and log sigma for the nth iterate
}
\description{
Gradient Vector for Truncated Normal Distribution with Censoring with Linear Equation Mean for J Independent Truncated Normals with Seprate Variance
}
