% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcensReg_hess.R
\name{tcensReg_hess}
\alias{tcensReg_hess}
\title{Hessian Matrix for Truncated Normal Distribution with Censoring with Linear Equation Mean}
\usage{
tcensReg_hess(theta, y, X, a = -Inf, v = NULL)
}
\arguments{
\item{theta}{Numeric vector numeric vector containing estimates of beta and log sigma}

\item{y}{Numeric vector with the observed truncated and censored outcomes}

\item{X}{Numeric design matrix}

\item{a}{Numeric scalar indicating the truncation value}

\item{v}{Numeric scalar indicating the censoring value}
}
\value{
Matrix of Hessian values for the nth iterate
}
\description{
Hessian Matrix for Truncated Normal Distribution with Censoring with Linear Equation Mean
}
