% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcensReg_llike.R
\name{tcensReg_llike}
\alias{tcensReg_llike}
\title{Log Likelihood for Truncated Normal Distribution with Censoring with Linear Equation Mean}
\usage{
tcensReg_llike(theta, y, X, a = -Inf, v = NULL, xi = NULL, b = Inf)
}
\arguments{
\item{theta}{Numeric vector containing estimates of \eqn{\beta} and \eqn{\log(\sigma)}}

\item{y}{Numeric vector with the observed truncated and censored outcomes}

\item{X}{Numeric design matrix}

\item{a}{Numeric scalar indicating the left truncation value. Default is -Inf
indicating no truncation}

\item{v}{Numeric scalar indicating the left censoring value. Default is NULL
indicating no censoring}

\item{xi}{Numeric scalar indicating the right censoring valye. Default is
NULL indicating to right censoring}

\item{b}{Numeric scalar indicating the right truncation value. Default is Inf
indicating no truncation}
}
\value{
Scalar value of the log-likelihood at the nth iterate
}
\description{
This is a supporting function that is used to calculate the log likelihood value at the nth
iteration of theta.
}
\details{
If \code{a} and/or \code{v} are not specified then the corresponding censored only, truncated only,
or gaussian log likelihood will be used. This function is called as part of
the Newton-Raphson algorithm in \code{tcensReg_newton} and in additional
optimization methods in \code{tcensReg_optim}.
}
\keyword{internal}
