% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcensReg_llike_sepvar.R
\name{tcensReg_llike_sepvar}
\alias{tcensReg_llike_sepvar}
\title{Log Likelihood for Model for J independent Truncated Normal Variables with Same Mean Structure but Different Variance}
\usage{
tcensReg_llike_sepvar(
  theta,
  y,
  X,
  group,
  a = -Inf,
  v = NULL,
  xi = NULL,
  b = Inf
)
}
\arguments{
\item{theta}{Numeric vector containing the values at the present iterate for the (p-1) fixed mean parameters and the J log sigma values}

\item{y}{Numeric vector containing the outcome}

\item{X}{Numeric design matrix}

\item{group}{Factor variable used to define the J groups}

\item{a}{Numeric scalar defining the common left-truncation value for each Truncated Normal}

\item{v}{Numeric scalar defining the common left-censoring value}

\item{xi}{Numeric scalar defining the common right-censoring value}

\item{b}{Numeric scalar defining the common right-truncation value for each Truncated Normal}
}
\value{
Scalar value of the log-likelihood at the nth-iterate
}
\description{
This function is part of the supporting function space to calculate the value of the log-likelihood at the nth iteration of theta.
For this parametrization we assume that there exists (p-1) linear parameters for the mean and J separate variance parameters.
}
\keyword{internal}
