% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcensReg_llike_sepvar_maxLik.R
\name{tcensReg_llike_sepvar_maxLik}
\alias{tcensReg_llike_sepvar_maxLik}
\title{Log Likelihood for Model for J independent Truncated Normal Variables with Same Mean Structure but Different Variance}
\usage{
tcensReg_llike_sepvar_maxLik(theta, y, X, group, left_trunc = -Inf, v = NULL)
}
\arguments{
\item{theta}{Numeric vector containing the values at the present iterate for the (p-1) fixed mean parameters and the J log sigma values}

\item{y}{Numeric vector containing the outcome}

\item{X}{Numeric design matrix}

\item{group}{Factor variable used to define the J groups}

\item{left_trunc}{Numeric scalar defining the common truncation value for each Truncated Normal}

\item{v}{Numeric scalar defining the common censoring value}
}
\value{
Scalar value of the log-likelihood at the nth-iterate
}
\description{
This function is part of the supporting function space to calculate the value of the log-likelihood at the nth iteration of theta.
For this parametrization we assume that there exists (p-1) linear parameters for the mean and J separate variance parameters.
}
\keyword{internal}
