% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcensReg.R
\name{tcensReg}
\alias{tcensReg}
\title{Regression Method for Truncated Normal Distribution with Censoring}
\usage{
tcensReg(formula, a = -Inf, v = NULL, data = sys.frame(sys.parent()),
  ...)
}
\arguments{
\item{formula}{Object of class \code{formula} which symbolically describes the model to be fit}

\item{a}{Numeric scalar indicating the truncation value. Initial value is -Inf indicating no truncation}

\item{v}{Numeric scalar indicating the censoring value. Initially set to NULL indicating no censoring}

\item{...}{Additional arguments from \code{\link{tcensReg_newton}} such as \code{max_iter}, \code{step_max}, or \code{epsilon}.}
}
\value{
Returns a list of final estimate of theta, total number of iterations performed, initial log-likelihood,
final log-likelihood, and estimated variance covariance matrix.
}
\description{
This function is used to find estimates from a linear equation assuming that the data is observed from a truncated
 distribution with left censoring. It uses analytically derived values of the gradient vector and hessian matrix to
 iteratively solve for the maximum likelihood using Newton-Raphson methods with line search. This function can also
 be used with censored only, truncated only, or uncensored and untruncated gaussian models.
}
