% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcensReg_hess.R
\name{tcensReg_hess}
\alias{tcensReg_hess}
\title{Hessian Matrix for Truncated Normal Distribution with Censoring with Linear Equation Mean}
\usage{
tcensReg_hess(theta, y, X, a = -Inf, v = NULL, xi = NULL, b = Inf)
}
\arguments{
\item{theta}{Numeric vector containing estimates of \eqn{\beta} and \eqn{\log(\sigma)}}

\item{y}{Numeric vector with the observed truncated and censored outcomes}

\item{X}{Numeric design matrix}

\item{a}{Numeric scalar indicating the left truncation value. Default is -Inf
indicating no truncation}

\item{v}{Numeric scalar indicating the left censoring value. Default is NULL
indicating no censoring}

\item{xi}{Numeric scalar indicating the right censoring valye. Default is
NULL indicating to right censoring}

\item{b}{Numeric scalar indicating the right truncation value. Default is Inf
indicating no truncation}
}
\value{
Matrix of Hessian values for the nth iterate
}
\description{
Hessian Matrix for Truncated Normal Distribution with Censoring with Linear Equation Mean
}
\keyword{internal}
